pyproject.toml
[project]
authors         = [{ name="Matthew Taylor", email="matthew@m-taylor.co.uk" }]
description     = "A Python script for generating OTP codes"
name            = "py_oathtool"
readme          = "README.md"
requires-python = ">=3.8"
version         = "1.2.0"

classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Intended Audience :: System Administrators",
  "License :: OSI Approved :: GNU General Public License v2 (GPLv2)",
  "Programming Language :: Python :: 3",
  "Topic :: Security",
  "Topic :: System :: Systems Administration",
  "Topic :: Utilities",
]

dependencies = [
  "oathtool",
  "pyaml",
]

[project.urls]
Homepage = "https://github.com/matalo33/py_oathtool"
Issues   = "https://github.com/matalo33/py_oathtool/issues"

[build-system]
build-backend = "setuptools.build_meta"
requires      = ["setuptools>=61.0"]

[project.scripts]
otp = "py_oathtool.otp:main"
